html/semantics/embedded-content/the-video-element/video-loading-lazy-in-viewport.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<video id="target" src="/media/A4.mp4?pipe=trickle(d2)" loading="lazy"></video>

<script>
  const video = document.getElementById("target");

  let windowLoadFired = false;

  async_test(t => {
    window.addEventListener("load", t.step_func(() => {
      windowLoadFired = true;
    }));

    video.addEventListener("loadstart", t.step_func(() => {
      assert_true(windowLoadFired, "Window load event should fire before loadstart for lazy video in viewport");
      t.done();
    }));

  }, "In-viewport video with loading='lazy' fires window load before loadstart");
</script>
